Add optional third parameter to constructor, allowing showing a specific tab
Fixes #1
Optional parameter can be either a number or a function. A function will be
called and expected to return a number. The number is used as the tab index
and supplied to showTab().

Current implementation is generic, extended versions of the class could be
made to take a more hard coded approach, this is up to the user.