Use location parameter in add(), allowing placement of a new tab and content
Fixes #2
Fixes #3
Currently, the parameter only works as a number that is treated as the
intended index of the new content. When location is specified, injectBefore
(or rather inject(element, 'before')) is used, making it impossible to insert
new content at the last index using location. Leaving the parameter undefined
(or anything else than a number) will place both at the end of their siblings.